+++ src/main/java/com/prueba/Agilesoft/entity/Usuario.java
@@ -22,7 +22,7 @@ public class Usuario {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     @Column(name = "idUsuario")
-    private Integer idUsuario;
+    private Long idUsuario;
 
     @Column(name = "nombre")
     private String nombre;
@@ -36,18 +36,18 @@ public class Usuario {
     public Usuario() {
     }
 
-    public Usuario(Integer idUsuario, String nombre, String username, String password) {
+    public Usuario(Long idUsuario, String nombre, String username, String password) {
         this.idUsuario = idUsuario;
         this.nombre = nombre;
         this.username = username;
         this.password = password;
     }
 
-    public Integer getIdUsuario() {
+    public Long getIdUsuario() {
         return idUsuario;
     }
 
-    public void setIdUsuario(Integer idUsuario) {
+    public void setIdUsuario(Long idUsuario) {
         this.idUsuario = idUsuario;
     }
